rabbitmq: Enable deploying rabbitmq with clustering when doing HA
Clustering enables us to not require shared storage, and also to have
active/active HA for rabbitmq.

The configuration of OpenStack services needs to be adapted so that the
services are aware of the multiple rabbitmq servers, and the services
will then connect to one server, and if that connection fails, will try
to connect to another one.

(cherry picked from commit 72f5b349bd9499a9b2f24fb3467dbd3d24f3e6e1)

Conflicts resolved:
      chef/cookbooks/rabbitmq/templates/default/rabbitmq.config.erb
      chef/data_bags/crowbar/template-rabbitmq.json
      chef/data_bags/crowbar/template-rabbitmq.schema