cinder: add nfs_mount_options configuration parameter
nfs_mount_options is a list of mount options that can be used when
cinder is mounting nfs shares for volumes. These options will be applied
on all volumes, so it's useful for situations where a share is no longer
listed in the nfs_shares list, but cinder still needs to manage volumes
living on it.