postgresql - reload: Graceful reload of configuration
This change splits the postgresql configuration file into two pieces.
One should contain the options that require a server restart (the main
postgresql.conf file) and that is sourcing the other file
(postgresql_reloadable.conf) which is the home for reloadable
parameters. This way we can use chef's notification mechanism to either
reload or restart the server.

Please consult postgresql's manual for reloadable configuration options.
If you want to be able to gracefuly reload a specific parameter, amend
that to `reloadable_parameters`