neutron: lbaasv2: optional haproxy when using f5
When people use F5 they might still want to have haproxy as a
non-default. This change will add a raw parameter
add_haproxy_as_non_default to the neutron barclamp/f5. When user sets
that value to true, HAProxy will be added as a non-default lbaasv2
driver to the neutron_lbaas.conf